feat: check statistics before delete devices in room

+++ admin_php/delete_fan.php
@@ -4,21 +4,28 @@ include '../database/config.php';
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
     $fanID = $_POST['fanID'];
 
-    // Insert device into device table
-    $fanSql = "DELETE FROM fan WHERE deviceID = $fanID";
+    // Delete from statistics table
+    $statisticsSql = "DELETE FROM statistics WHERE deviceID = $fanID";
 
-    if ($conn->query($fanSql) === TRUE) {
-        // Insert corresponding entry into light table
-        $deviceSql = "DELETE FROM device WHERE deviceID = $fanID";
+    if ($conn->query($statisticsSql) === TRUE) {
+        // Delete from fan table
+        $fanSql = "DELETE FROM fan WHERE deviceID = $fanID";
 
-        if ($conn->query($deviceSql) === TRUE) {
-            header("Location: " . $_SERVER["HTTP_REFERER"]); 
-            exit();
+        if ($conn->query($fanSql) === TRUE) {
+            // Delete from device table
+            $deviceSql = "DELETE FROM device WHERE deviceID = $fanID";
+
+            if ($conn->query($deviceSql) === TRUE) {
+                header("Location: " . $_SERVER["HTTP_REFERER"]); 
+                exit();
+            } else {
+                echo "Error deleting from device table: " . $conn->error;
+            }
         } else {
-            echo "Error inserting into light table: " . $conn->error;
+            echo "Error deleting from fan table: " . $conn->error;
         }
     } else {
-        echo "Error inserting into device table: " . $conn->error;
+        echo "Error deleting from statistics table: " . $conn->error;
     }
 
     $conn->close();
