+++ admin_php/delete_toy.php
@@ -4,21 +4,28 @@ include '../database/config.php';
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
     $toyID = $_POST['toyID'];
 
-    // Insert device into device table
-    $toySql = "DELETE FROM robot_toy WHERE deviceID = $toyID";
+    // Delete from statistics table
+    $statisticsSql = "DELETE FROM statistics WHERE deviceID = $toyID";
 
-    if ($conn->query($toySql) === TRUE) {
-        // Insert corresponding entry into light table
-        $deviceSql = "DELETE FROM device WHERE deviceID = $toyID";
+    if ($conn->query($statisticsSql) === TRUE) {
+        // Delete from robot_toy table
+        $toySql = "DELETE FROM robot_toy WHERE deviceID = $toyID";
 
-        if ($conn->query($deviceSql) === TRUE) {
-            header("Location: " . $_SERVER["HTTP_REFERER"]); 
-            exit();
+        if ($conn->query($toySql) === TRUE) {
+            // Delete from device table
+            $deviceSql = "DELETE FROM device WHERE deviceID = $toyID";
+
+            if ($conn->query($deviceSql) === TRUE) {
+                header("Location: " . $_SERVER["HTTP_REFERER"]); 
+                exit();
+            } else {
+                echo "Error deleting from device table: " . $conn->error;
+            }
         } else {
-            echo "Error inserting into light table: " . $conn->error;
+            echo "Error deleting from robot_toy table: " . $conn->error;
         }
     } else {
-        echo "Error inserting into device table: " . $conn->error;
+        echo "Error deleting from statistics table: " . $conn->error;
     }
 
     $conn->close();
