+++ admin_php/delete_vacuumcleaner.php
@@ -4,21 +4,28 @@ include '../database/config.php';
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
     $vacuumID = $_POST['vacuumID'];
 
-    // Insert device into device table
-    $vacuumSql = "DELETE FROM robot_vacum_cleaner WHERE deviceID = $vacuumID";
+    // Delete from statistics table
+    $statisticsSql = "DELETE FROM statistics WHERE deviceID = $vacuumID";
 
-    if ($conn->query($vacuumSql) === TRUE) {
-        // Insert corresponding entry into light table
-        $deviceSql = "DELETE FROM device WHERE deviceID = $vacuumID";
+    if ($conn->query($statisticsSql) === TRUE) {
+        // Delete from robot_vacum_cleaner table
+        $vacuumSql = "DELETE FROM robot_vacum_cleaner WHERE deviceID = $vacuumID";
 
-        if ($conn->query($deviceSql) === TRUE) {
-            header("Location: " . $_SERVER["HTTP_REFERER"]); 
-            exit();
+        if ($conn->query($vacuumSql) === TRUE) {
+            // Delete from device table
+            $deviceSql = "DELETE FROM device WHERE deviceID = $vacuumID";
+
+            if ($conn->query($deviceSql) === TRUE) {
+                header("Location: " . $_SERVER["HTTP_REFERER"]); 
+                exit();
+            } else {
+                echo "Error deleting from device table: " . $conn->error;
+            }
         } else {
-            echo "Error inserting into light table: " . $conn->error;
+            echo "Error deleting from robot_vacum_cleaner table: " . $conn->error;
         }
     } else {
-        echo "Error inserting into device table: " . $conn->error;
+        echo "Error deleting from statistics table: " . $conn->error;
     }
 
     $conn->close();
